+++ src/main/java/com/second/kirby/controller/UserController.java
@@ -2,8 +2,8 @@ package com.second.kirby.controller;
 
 import com.second.kirby.domain.User;
 import com.second.kirby.dto.ResponseDto;
-import com.second.kirby.dto.request.PasswordChangeRequest;
-import com.second.kirby.dto.request.ProfileUpdateRequest;
+import com.second.kirby.dto.request.user.PasswordChangeRequest;
+import com.second.kirby.dto.request.user.ProfileUpdateRequest;
 import com.second.kirby.exception.BusinessException;
 import com.second.kirby.exception.ResponseCode;
 import com.second.kirby.service.UserService;
